+++ src/store/userSlice.js
@@ -47,13 +47,10 @@ export const logoutUser = () => {
 		window.localStorage.removeItem("token");
 	};
 };
-export const createUser = (userDetails, isAdmin = false) => {
+export const createUser = (userDetails) => {
 	return async (dispatch) => {
 		try {
-			const { data: user } = await axios.post("/api/users", {
-				...userDetails,
-				isAdmin,
-			});
+			const { data: user } = await axios.post("/api/users", userDetails);
 			if (user) {
                 attemptPsswordLogin({username: userDetails.username, password: userDetails.password})(dispatch);
 			}
